#1551 test: update upload test + TestFilesystemTrait
